GraphService double +Updated GraphService class functions to double instead of float

+++ src/game/world/FlatGraph.h
@@ -14,10 +14,10 @@ public:
     virtual ~FlatGraph();
 
     // Implementation vom GraphService Interface
-    virtual float heightFunction(float valueX);
-    virtual float heightFunctionDerivation(float valueX);
-    virtual float depthFunction(float valueZ);
-    virtual float depthFunctionDerivation(float valueZ);
+    virtual double heightFunction(float valueX);
+    virtual double heightFunctionDerivation(float valueX);
+    virtual double depthFunction(float valueZ);
+    virtual double depthFunctionDerivation(float valueZ);
 };
